requirements.txt
google-api-core==1.31.5
google-auth==1.35.0
google-auth-oauthlib==0.5.1
google-cloud-bigquery==2.18.0
google-cloud-bigquery-storage==2.13.1
google-cloud-core==1.7.2
google-cloud-storage==1.38.0
google-crc32c==1.3.0
google-resumable-media==1.3.3
googleapis-common-protos==1.56.1
numpy==1.20.2
pandas==1.2.4
pandas-gbq==0.17.5
SQLAlchemy==1.4.13
PyMySQL==1.0.2
jupyterlab==3.0.14
notebook==6.3.0
jupyterthemes==0.20.0
pyyaml==5.4.1
Flask==2.1.2
Flask-WTF==1.0.1
Flask-Bootstrap4==4.0.2